requirement.txt
transformers
torch
umap-learn
joblib
sklearn
gensim
PyPDF2
pandas
nltk
